Implementing table schemas
This PR adds the ability to prefix tables with a database schema.
The API has supported this feature for a _long_ time but for some
reason the feature was silently removed in a refactoring (woops).

Test coverage is still a bit lacking. This is adding a single test
for each backend to make sure it roughly works. And most importantly
tests that non-schema prefixed actions don't break existing tests :)