Making diesel integration code backend configurable (#60) (fixes #59)
* Making diesel integration code backend configurable

This is a first attempt at solving the problem that diesel doesn't
tell us what kind of migration is being run, just demanding an Sql
back from us. We could probably fix this in a nicer way on diesel's
side (i.e. changing the API) but that is a lot more work that just
getting it to work now.

This change assumes that `barrel` is only being compiled with a single
feature, meaning that whenever adding `barrel` to your `diesel` setup,
you can only select one backend.

### Additional commits

* Adjusting travis-ci builds to not include unbuildable feature sets
* Fixing the format string issue and cleaning up code
* Updating the diesel guide to mention how to switch SQL flavour
* Fixing sqlite3 feature name
* Adjusting debug printing to new output