Removing the `debugging_types` test
This test was meant to make sure that the internal layout and
representation of a type doesn't change. Since we aren't actually
stable yet (i.e. `1.0.0`) this isn't really required or a good
idea. There've been a few breakages recently because of this test and
it's getting annoying to debug.

So we're removing it for now. Maybe at some point in the future we can
add another type of tests that does a similar thing again...