Re-implementing foreign reference API
The previous builder API was very suboptimal because it required
database-specific string escaping on the side of the user. Instead
`barrel` now takes a table to reference as a string, then either a
single column name or a vector of columns.
This is currently still "stringly-typed" but good enough for now!

It also adds a `foreign_schema` function which can be passed a schema
that is _different_ from the current table schema.